Renaming viscous laws (in order of a future 2D compatibility)
