+++ applications/StructuralMechanicsApplication/custom_constitutive/viscous_generalized_kelvin.cpp
@@ -13,7 +13,7 @@
 // System includes
 
 // Project includes
-#include "custom_constitutive/viscous_generalized_kelvin_3d.h"
+#include "custom_constitutive/viscous_generalized_kelvin.h"
 #include "structural_mechanics_application_variables.h"
 
 namespace Kratos
