+++ applications/StructuralMechanicsApplication/custom_constitutive/viscous_generalized_maxwell.cpp
@@ -18,7 +18,7 @@
 #include "includes/checks.h"
 #include "utilities/math_utils.h"
 #include "structural_mechanics_application_variables.h"
-#include "custom_constitutive/viscous_generalized_maxwell_3d.h"
+#include "custom_constitutive/viscous_generalized_maxwell.h"
 
 namespace Kratos
 {
