+++ applications/StructuralMechanicsApplication/structural_mechanics_application.h
@@ -106,9 +106,9 @@
 #include "custom_constitutive/generic_small_strain_isotropic_plasticity.h"
 #include "custom_constitutive/generic_small_strain_isotropic_damage.h"
 #include "custom_constitutive/small_strain_isotropic_damage_factory.h"
-#include "custom_constitutive/viscous_generalized_kelvin_3d.h"
+#include "custom_constitutive/viscous_generalized_kelvin.h"
+#include "custom_constitutive/viscous_generalized_maxwell.h"
 #include "custom_constitutive/generic_small_strain_viscoplasticity_3d.h"
-#include "custom_constitutive/viscous_generalized_maxwell_3d.h"
 
 // Integrators
 #include "custom_constitutive/constitutive_laws_integrators/generic_constitutive_law_integrator_damage.h"
