+++ applications/StructuralMechanicsApplication/tests/cpp_tests/test_kelvin_constitutive_law.cpp
@@ -20,7 +20,7 @@
 // Application includes
 
 // Constitutive law
-#include "custom_constitutive/viscous_generalized_kelvin_3d.h"
+#include "custom_constitutive/viscous_generalized_kelvin.h"
 #include "includes/model_part.h"
 #include "geometries/tetrahedra_3d_4.h"
 #include "structural_mechanics_application_variables.h"
