+++ applications/StructuralMechanicsApplication/tests/cpp_tests/test_maxwell_constitutive_law.cpp
@@ -20,7 +20,7 @@
 // Application includes
 
 // Constitutive law
-#include "custom_constitutive/viscous_generalized_maxwell_3d.h"
+#include "custom_constitutive/viscous_generalized_maxwell.h"
 #include "includes/model_part.h"
 #include "geometries/tetrahedra_3d_4.h"
 #include "structural_mechanics_application_variables.h"
